Optimize Dijkstra's algorithm on day 15
For 250,000 nodes, the running time decrease from 4299s to 12s.
The main difference is that instead of populating the queue with all nodes at the start, initially the queue is empty, and we (1) only add nodes as they are reached (2) take a node out when it is chosen to be the next min value. The benefit is at the step of choosing the min value from the queue, the data size that needs to be searched is much smaller, yielding a faster result.
Ref: https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm#Practical_optimizations_and_infinite_graphs
Next steps:
1. Implement a priority queue (Currently using hash)
2. Refactor...